Move `VisConfigProvider` to `vis-packs` folder
